+++ tests/Feature/Framework/StaticFacade/Module/Domain/GreeterInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\StaticFacade\Module\Domain;
+
+interface GreeterInterface
+{
+    public function greet(string $name): string;
+}