+++ tests/Feature/Framework/StaticFacade/Module/Facade.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Framework\StaticFacade\Module;
+
+use Gacela\Framework\AbstractFacade;
+
+/**
+ * @method Factory getFactory()
+ * @method static Factory factory()
+ */
+final class Facade extends AbstractFacade
+{
+    public static function informalGreet(string $name): string
+    {
+        return self::factory()
+            ->createInformatGreeter()
+            ->greet($name);
+    }
+
+    public function formalGreet(string $name): string
+    {
+        return $this->getFactory()
+            ->createFormatGreeter()
+            ->greet($name);
+    }
+}