+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Unreleased
+
+- Allow using static facade methods
+
 ### 1.0.1
 ### 2023-03-12